Remove hydra
