+++ ultralytics/yolo/engine/model.py
@@ -12,15 +12,18 @@ from ultralytics.yolo.utils.torch_utils import guess_task_from_head, smart_infer
 
 # Map head to model, trainer, validator, and predictor classes
 MODEL_MAP = {
-    "classify": [
-        ClassificationModel, 'yolo.TYPE.classify.ClassificationTrainer', 'yolo.TYPE.classify.ClassificationValidator',
-        'yolo.TYPE.classify.ClassificationPredictor'],
-    "detect": [
-        DetectionModel, 'yolo.TYPE.detect.DetectionTrainer', 'yolo.TYPE.detect.DetectionValidator',
-        'yolo.TYPE.detect.DetectionPredictor'],
-    "segment": [
-        SegmentationModel, 'yolo.TYPE.segment.SegmentationTrainer', 'yolo.TYPE.segment.SegmentationValidator',
-        'yolo.TYPE.segment.SegmentationPredictor']}
+    "classify": [ClassificationModel,
+                 'yolo.TYPE.classify.ClassificationTrainer',
+                 'yolo.TYPE.classify.ClassificationValidator',
+                 'yolo.TYPE.classify.ClassificationPredictor'],
+    "detect": [DetectionModel,
+               'yolo.TYPE.detect.DetectionTrainer',
+               'yolo.TYPE.detect.DetectionValidator',
+               'yolo.TYPE.detect.DetectionPredictor'],
+    "segment": [SegmentationModel,
+                'yolo.TYPE.segment.SegmentationTrainer',
+                'yolo.TYPE.segment.SegmentationValidator',
+                'yolo.TYPE.segment.SegmentationPredictor']}
 
 
 class YOLO:
