+++ ultralytics/yolo/v8/detect/predict.py
@@ -81,7 +81,7 @@ class DetectionPredictor(BasePredictor):
         return log_string
 
 
-@hydra.main(version_base=None, config_path=str(DEFAULT_CONFIG.parent), config_name=DEFAULT_CONFIG.name)
+
 def predict(cfg):
     cfg.model = cfg.model or "yolov8n.pt"
     cfg.source = cfg.source if cfg.source is not None else ROOT / "assets" if is_git_directory() \
