+++ ultralytics/yolo/v8/segment/val.py
@@ -243,7 +243,6 @@ class SegmentationValidator(DetectionValidator):
         return stats
 
 
-@hydra.main(version_base=None, config_path=str(DEFAULT_CONFIG.parent), config_name=DEFAULT_CONFIG.name)
 def val(cfg):
     cfg.data = cfg.data or "coco128-seg.yaml"
     validator = SegmentationValidator(args=cfg)
